be/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
